backend-app/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==10.4.0
pytz==2024.1